+++ test/raxx/basic_auth_test.exs
@@ -1,0 +1,7 @@
+defmodule Raxx.BasicAuthTest do
+  use ExUnit.Case, async: true
+  import Raxx
+  import Raxx.BasicAuth
+
+  doctest Raxx.BasicAuth
+end