+++ CHANGELOG.md
@@ -4,11 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## master
+## [0.14.13](https://github.com/CrowdHailer/raxx/tree/0.14.13) - 2018-03-19
 
 ### Added
+
 - Secure sessions in signed cookies by using `Raxx.Session.SignedCookie`.
-- `Raxx.BasicAuth` for setting and verifying credentials using the Basic authentication scheme. 
+- `Raxx.BasicAuth` for setting and verifying credentials using the Basic authentication scheme.
 
 ## [0.14.12](https://github.com/CrowdHailer/raxx/tree/0.14.12) - 2018-03-14